Refactor kinematic bicycle model to use float instead of double
- Updated kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_A.hpp to use float for matrix types and values in make function.
- Changed kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_C.hpp to utilize float for diagonal matrix type.
- Modified kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_measurement_function.hpp to replace double with float in state and output types.
- Adjusted kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_measurement_function_jacobian.hpp to use float for Jacobian calculations.
- Altered kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_parameter.hpp to define parameters as float.
- Updated kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_state_function.hpp to use float for state calculations.
- Changed kinematic_bicycle_model_op_en_nonlinear_mpc_ekf_state_function_jacobian.hpp to utilize float for Jacobian calculations.
- Refactored kinematic_bicycle_model_op_en_oe_measurement_function.hpp to use float for measurement calculations.
- Updated kinematic_bicycle_model_op_en_oe_state_function.hpp to define state calculations using float.
- Adjusted kinematic_bicycle_model_op_en_oe_state_jacobian_u.hpp to use float for state Jacobian calculations.
- Modified kinematic_bicycle_model_op_en_oe_state_jacobian_x.hpp to utilize float for state Jacobian calculations.
- Changed kinematic_bicycle_model_op_en_parameter.hpp to define parameters as float.

+++ include/Tester/mpc/nonlinear_mpc_optimization_engine/kinematic_bicycle_model_op_en_cost_matrices_U_min.hpp
@@ -7,15 +7,12 @@ namespace kinematic_bicycle_model_op_en_cost_matrices_U_min {
 
 using namespace PythonNumpy;
 
-using type = DenseMatrix_Type<double, 2, 1>;
+using type = DenseMatrix_Type<float, 2, 1>;
 
 inline auto make(void) -> type {
 
-  return make_DenseMatrix<2, 1>(
-    static_cast<double>(1.0),
-    static_cast<double>(1.0)
-  );
-
+  return make_DenseMatrix<2, 1>(static_cast<float>(1.0),
+                                static_cast<float>(1.0));
 }
 
 } // namespace kinematic_bicycle_model_op_en_cost_matrices_U_min
